break Fast Api into separate modulr
